Refactor: Move I/O functions from utils to new io module
- Create new mdx2/io.py with file I/O functions (nxload, nxsave, loadobj, saveobj, _patch_virtualfields)
- Keep numerical/computational functions in mdx2/utils.py (interpolation, slice_sections)
- Update all imports across 12 command-line tools, core modules, tests, and example notebooks
- All tests pass successfully (34/34)
