+++ mdx2/command_line/import_geometry.py
@@ -10,7 +10,7 @@ from simple_parsing import ArgumentParser, field
 
 import mdx2.geometry as geom
 from mdx2.command_line import with_logging
-from mdx2.utils import saveobj
+from mdx2.io import saveobj
 
 
 @dataclass
