+++ mdx2/command_line/mask_peaks.py
@@ -11,7 +11,7 @@ from simple_parsing import ArgumentParser, field  # pip install simple-parsing
 
 from mdx2.command_line import log_parallel_backend, with_logging
 from mdx2.geometry import GridData
-from mdx2.utils import loadobj, saveobj
+from mdx2.io import loadobj, saveobj
 
 
 @dataclass
